% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causality_predVAR.R, R/causality_predVAR_v0.R
\name{causality_resVAR}
\alias{causality_resVAR}
\title{Out-of-sample Tests of Granger Causality using Restricted Vector Autoregression}
\usage{
causality_resVAR(y, p = NULL, cause = NULL, B = 100, test = 0.3, ...)

causality_resVAR(y, p = NULL, cause = NULL, B = 100, test = 0.3, ...)
}
\arguments{
\item{y}{data frame or \code{ts} object for estimating VAR(\eqn{p}).}

\item{p}{an integer specifying the order \eqn{p} in VAR.
By default (if \code{p} is not specified),
\eqn{p} is selected based on the information criterion
\code{ic} (see \code{...} arguments; default \code{ic} is AIC).}

\item{cause}{name of the cause variable. If not specified, the first variable in
\code{y} is treated as the cause, and second -- as the dependent variable.}

\item{B}{number of bootstrap replications. Default is 100.}

\item{test}{a numeric value specifying the size of the testing set. If \code{test} < 1,
the value is treated as proportion of the sample size to be used as the testing set.
Otherwise, \code{test} is rounded and \code{test} values are used as the testing set.
Default is 0.3, which means that 30\% of the sample are used for calculating
out-of-sample errors. The testing set is always at the end of the time series.}

\item{...}{other arguments passed to the function for VAR estimation.
The arguments include \code{lag.restrict} that is used to remove a number of first lags
in the cause variable from consideration (use restricted VAR to avoid testing for short-term causality);
default \code{lag.restrict = 0L}, i.e., no restrictions.
Other possible arguments are as in the \code{\link[vars]{VAR}} function.
Also see Details and Examples.}
}
\value{
A list containing the following elements:
\item{MSEt}{observed value of the MSEt statistic.}
\item{MSEt_p}{bootstrapped \eqn{p}-value of the MSEt test.}
\item{MSEt_p_asympt}{asymptotic \eqn{p}-value of the MSEt test,
based on the left tail of the t distribution.}
\item{MSEcor}{observed value of the MSEcor statistic.}
\item{MSEcor_p}{bootstrapped \eqn{p}-value of the MSEcor test.}
\item{MSEcor_p_asympt}{asymptotic \eqn{p}-value of the MSEcor test,
based on the left tail of the t distribution.}
\item{Md}{observed value of the Md statistic.}
\item{Md_p}{bootstrapped \eqn{p}-value of the Md test.}
\item{p}{order \eqn{p} of the VAR(\eqn{p}) model.}

A list containing the following elements:
\item{MSEt}{observed value of the MSEt statistic.}
\item{MSEt_p}{bootstrapped \eqn{p}-value of the MSEt test.}
\item{MSEt_p_asympt}{asymptotic \eqn{p}-value of the MSEt test,
based on the left tail of the t distribution.}
\item{MSEcor}{observed value of the MSEcor statistic.}
\item{MSEcor_p}{bootstrapped \eqn{p}-value of the MSEcor test.}
\item{MSEcor_p_asympt}{asymptotic \eqn{p}-value of the MSEcor test,
based on the left tail of the t distribution.}
\item{Md}{observed value of the Md statistic.}
\item{Md_p}{bootstrapped \eqn{p}-value of the Md test.}
\item{p}{order \eqn{p} of the VAR(\eqn{p}) model.}
}
\description{
Test for Granger causality using out-of-sample prediction errors from a vector
autoregression (VAR), where the original VAR can be a restricted VAR (see Details).
The tests include the MSE-t approach \insertCite{McCracken_2007}{funtimes},
MSE-correlation test as in Chapter 9.3 of \insertCite{Granger_Newbold_2016;textual}{funtimes},
and difference of squared VAR prediction errors (Md statistic) when not using the cause
variable and with the cause variable.
Bootstrap is used to empirically derive distributions of the statistics.

Test for Granger causality using out-of-sample prediction errors from a vector
autoregression (VAR), where the original VAR can be a restricted VAR (see Details).
The tests include the MSE-t approach \insertCite{McCracken_2007}{funtimes},
MSE-correlation test as in Chapter 9.3 of \insertCite{Granger_Newbold_2016;textual}{funtimes},
and difference of squared VAR prediction errors (Md statistic) when not using the cause
variable and with the cause variable.
Bootstrap is used to empirically derive distributions of the statistics.
}
\details{
The arguments specified in \code{...} are passed to the \code{\link[vars]{VAR}} function.
Additionally, \code{lag.restrict} can be specified to remove short-term lags from
consideration (\code{lag.restrict} is not an option in the original package \code{vars}).
Note that if \code{p} is specified, \code{lag.restrict} must be smaller
than \code{p} otherwise the default \code{lag.restrict = 0} will be used.
If \code{lag.max} is specified instead of \code{p}, VAR orders
\code{lag.restrict} + 1, \dots, \code{lag.max} will be considered using the training data
and the order \eqn{p} will be automatically selected according to the information criterion
(by default, AIC).

In the current implementation, the bootstrapped \eqn{p}-value is calculated using equation 4.10 of
\insertCite{Davison_Hinkley_1997;textual}{funtimes}: \code{p.value} = (1 + \eqn{n}) / (\code{B} + 1),
where \eqn{n} is number of bootstrapped statistics greater or equal to 0.

The arguments specified in \code{...} are passed to the \code{\link[vars]{VAR}} function.
Additionally, \code{lag.restrict} can be specified to remove short-term lags from
consideration (\code{lag.restrict} is not an option in the original package \code{vars}).
Note that if \code{p} is specified, \code{lag.restrict} must be smaller
than \code{p} otherwise the default \code{lag.restrict = 0} will be used.
If \code{lag.max} is specified instead of \code{p}, VAR orders
\code{lag.restrict} + 1, \dots, \code{lag.max} will be considered using the training data
and the order \eqn{p} will be automatically selected according to the information criterion
(by default, AIC).

In the current implementation, the bootstrapped \eqn{p}-value is calculated using equation 4.10 of
\insertCite{Davison_Hinkley_1997;textual}{funtimes}: \code{p.value} = (1 + \eqn{n}) / (\code{B} + 1),
where \eqn{n} is number of bootstrapped statistics greater or equal to 0.
}
\examples{
\dontrun{
cores <- parallel::detectCores()
cl <- parallel::makeCluster(cores)
loadedfunc <- lsf.str()
parallel::clusterExport(cl, varlist = as.list(loadedfunc), envir = environment())
# Example 1: Canada time series (ts object)
Canada <- vars::Canada
causality_resVAR(Canada[,1:2], cause = "e", lag.max = 5, cl = cl)
causality_resVAR(Canada[,1:2], cause = "e", lag.restrict = 3, cl = cl, lag.max = 15)

# Example 2: Box & Jenkins time series of sales and a leading indicator, see ?BJsales
D <- cbind(BJsales.lead, BJsales)
causality_resVAR(D, cause = "BJsales.lead", lag.max = 5)
causality_resVAR(D, cause = "BJsales.lead", lag.restrict = 3, p = 5, B = 100)
}

# Example 1: Canada time series (ts object)
Canada <- vars::Canada
causality_resVAR(Canada[,1:2], cause = "e", lag.restrict = 3, lag.max = 10)

# Example 2: Box & Jenkins time series of sales and a leading indicator, see ?BJsales
D <- cbind(BJsales.lead, BJsales)
causality_resVAR(D, cause = "BJsales.lead", lag.restrict = 3, lag.max = 10)
causality_resVAR(D, cause = "BJsales.lead", lag.restrict = 3, p = 5, B = 100)


}
\references{
\insertAllCited{}

\insertAllCited{}
}
\author{
Vyacheslav Lyubchich

Vyacheslav Lyubchich
}
\keyword{causality}
\keyword{htest}
\keyword{ts}
