% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CExpandSlideCluster.R
\name{CExpandSlideCluster}
\alias{CExpandSlideCluster}
\title{Slide-Level Time Series Cluster Expansion}
\usage{
CExpandSlideCluster(u, Xuncl, Alpha, Beta, Delta, Theta)
}
\arguments{
\item{u}{a time series vector -- a seed to expand the cluster.}

\item{Xuncl}{a time series vector (of the same length as \code{u}) or a matrix 
(time series in columns) containing unclustered time series.}

\item{Alpha}{lower limit of the time series domain.}

\item{Beta}{upper limit of the time series domain.}

\item{Delta}{closeness parameter, a real value in \eqn{[0,1]}.}

\item{Theta}{connectivity parameter, a real value in \eqn{[0,1]}.}
}
\value{
A vector of logical values indicating which time series in \code{Xuncl} 
should be included in the slide-level cluster with \code{u}.
}
\description{
This is an auxiliary function to expand a slide-level time series cluster, 
based on \insertCite{Ciampi_etal_2010;textual}{funtimes}.
}
\examples{
set.seed(123)
u <- rnorm(10)
Xuncl <- matrix(rt(50, 5), 10, 5)
Alpha <- min(cbind(u, Xuncl))
Beta <- max(cbind(u, Xuncl))
CExpandSlideCluster(u, Xuncl, Alpha, Beta, Delta = 0.15, Theta = 0.8)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{CSlideCluster}}, \code{\link{CWindowCluster}}, 
and \code{\link{BICC}}
}
\author{
Vyacheslav Lyubchich
}
\keyword{cluster,}
\keyword{trend}
\keyword{ts,}
