% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CExpandWindowCluster.R
\name{CExpandWindowCluster}
\alias{CExpandWindowCluster}
\title{Window-Level Time Series Cluster Expansion}
\usage{
CExpandWindowCluster(e, Euncl)
}
\arguments{
\item{e}{a vector of logical values identifying which time series among \code{Euncl} 
were clustered together with \code{e} over at least \code{w*Epsilon} slides 
within a window; see Definition 7 by \insertCite{Ciampi_etal_2010;textual}{funtimes}. 
This is a seed for window-level clustering.}

\item{Euncl}{a square matrix identifying the binary window cluster relation 
for yet unclustered time series.}
}
\value{
A vector of logical values indicating which time series in \code{Euncl} 
should be included in the window-level cluster with \code{e}.
}
\description{
This is an auxiliary function to expand a window-level time series cluster, 
based on \insertCite{Ciampi_etal_2010;textual}{funtimes}.
}
\examples{
set.seed(123)
e <- sample(c(TRUE, FALSE), 5, replace = TRUE)
Euncl <- matrix(sample(c(TRUE, FALSE), 5, replace = TRUE), 5, 5)
CExpandWindowCluster(e, Euncl)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{CSlideCluster}}, \code{\link{CWindowCluster}}, 
and \code{\link{BICC}}
}
\author{
Vyacheslav Lyubchich
}
\keyword{cluster}
\keyword{trend}
\keyword{ts}
