% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change-point-boot_DOB.R
\name{change.point.boot}
\alias{change.point.boot}
\title{Change Point Detection in Autoregressive Time Series}
\usage{
change.point.boot(y, a.order, alternatives = c("two-sided", "greater",
  "lesser", "temporary"), crit.type = c("asymptotic", "bootstrap"),
  num.bootstrap = 1000)
}
\arguments{
\item{y}{A vector that contains univariate time series observations. Missing values are not allowed.}

\item{a.order}{Order of the autoregressive model which must be a nonnegative integer number.}

\item{alternatives}{A string parameter that specifies a type of the test (i.e., "two-sided", "greater", "lesser", and "temporary").  The option "temporary" examines
the temporary change in one of the parameters (for details see \insertCite{Gombay_2008}{funtimes})}

\item{crit.type}{Method of obtaining critical values: "asymptotic" (default) or "bootstrap".}

\item{num.bootstrap}{Number of bootstrap replications if crit.type="bootstrap". Default number is 1,000.}
}
\value{
A list with the following components:
\item{index}{Points of change for each parameter. The value of the \code{"alternatives"}
 determines the return: 
"temporary - returns max, min and abs.max points;
"greater" - returns max points;
"lesser" -  returns min points;
"two-sided" - returns abs.max. }
\item{stats}{Test statistic values for change points in: mean, var, phi.}
\item{p.values}{\code{p-value} of the change point test.}
}
\description{
The function detects change points in autoregressive (AR) models for time series. Changes can be detected in any
of $p+2$ (mean, var, phi) autoregressive parameters where $p$ is the order of the AR model. The test statistic is based on 
the efficient score vector (See \insertCite{Gombay_2008}{funtimes}).
}
\details{
The function allows for
testing for a temporary change and for a change in a specific model parameters. 
Critical values can be estimated via asymptotic distribution \code{"asymptotic"} (i.e., the default option) or via sieve bootstrap \code{"bootstrap"}. 
The function employs internal function \code{change.point} and 
sieve bootstrap \code{change.point.sieve} function.
}
\examples{

#Example 1:
#Simulate some time series:
series_1 = arima.sim(n = 100, list(order = c(2,0,0), ar = c(-0.7, -0.1)))
series_2 = arima.sim(n = 200, list(order = c(2,0,0), ar = c(0.1, -0.6)))
main_series = c(series_1, series_2)

result11 = change.point.boot(series_1, 2, "two-sided")
result11 #== No change point ===#

result12 = change.point.boot(main_series, 2, "two-sided")
result12  #=== One change at phi values ===#

result13 = change.point.boot(main_series, 2, "two-sided", "bootstrap")
result13  #=== One change at phi values ===#



#Example 2:
#From the package 'Ecdat' consider a time series with annual world number of victims of terrorism in the US
#from 1970 till 2016:

c.data = (Ecdat::terrorism['nkill.us'])
nkill.us.ts<-ts(c.data,start=(1970), end=(2016), frequency(47))

#Now let's perform a change point detection with one sided tests:

change.point.boot(nkill.us.ts, 0, "lesser")
change.point.boot(nkill.us.ts, 0, "greater")
nkill.us.ts[32]
year=1970+31
print(year)
plot(nkill.us.ts)

#In both cases we find that the change point is located at the position 31 or 32. We can examine it further 

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{mcusum.test}} for change point test for regression
\code{\link[Ecdat]{terrorism}} dataset that was used for the Example 2
}
\author{
Poli Nemkova, Dorcas Ofori-Boateng, Yulia R. Gel
}
\keyword{series,}
\keyword{time}
\keyword{ts}
