% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNeighbor.R
\name{CNeighbor}
\alias{CNeighbor}
\title{Neighborhood of Time Series}
\usage{
CNeighbor(Bu, Bv, Alpha, Beta, Delta, Theta)
}
\arguments{
\item{Bu}{a time series vector for which the neighborhood is investigated.}

\item{Bv}{a time series vector (of the same length as \code{Bu}) or a matrix 
(time series in columns) containing potential neighbors.}

\item{Alpha}{lower limit of the time series domain.}

\item{Beta}{upper limit of the time series domain.}

\item{Delta}{closeness parameter, a real value in \eqn{[0,1]}.}

\item{Theta}{connectivity parameter, a real value in \eqn{[0,1]}.}
}
\value{
A vector of logical values indicating which time series in \code{Bv} 
are \eqn{E^\theta_\delta}-neighbors of \code{Bu}.
}
\description{
This is an auxiliary function to identify which time series in \code{Bv} 
are \eqn{E^\theta_\delta}-neighbors of \code{Bu}, based on Definition 2 
by \insertCite{Ciampi_etal_2010;textual}{funtimes}.
}
\examples{
Bu <- rnorm(10)
Bv <- rnorm(10)
Alpha <- min(c(Bu, Bv))
Beta <- max(c(Bu, Bv))
CNeighbor(Bu, Bv, Alpha, Beta, Delta = 0.5, Theta = 0.8)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{CSlideCluster}}, \code{\link{CWindowCluster}}, 
and \code{\link{BICC}}
}
\author{
Vyacheslav Lyubchich
}
\keyword{cluster,}
\keyword{trend}
\keyword{ts,}
