% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funtimes-defunct.R, R/i.tails.R,
%   R/mcusum.test.R, R/notrend.test.R, R/q.tails.R, R/sync.cluster.R,
%   R/sync.test.R, R/wavk.test.R
\name{funtimes-defunct}
\alias{funtimes-defunct}
\alias{i.tails}
\alias{mcusum.test}
\alias{notrend.test}
\alias{q.tails}
\alias{sync.cluster}
\alias{sync.test}
\alias{wavk.test}
\title{Defunct functions in package \pkg{funtimes}.}
\usage{
i.tails(...)

mcusum.test(...)

notrend.test(...)

q.tails(...)

sync.cluster(...)

sync.test(...)

wavk.test(...)
}
\description{
The functions listed below are defunct and no longer available.
}
\section{\code{i.tails}}{

For \code{i.tails}, use \code{\link{tails_i}}.
}

\section{\code{mcusum.test}}{

For \code{mcusum.test}, use \code{\link{mcusum_test}}.
}

\section{\code{notrend.test}}{

For \code{notrend.test}, use \code{\link{notrend_test}}.
}

\section{\code{q.tails}}{

For \code{q.tails}, use \code{\link{tails_q}}.
}

\section{\code{sync.cluster}}{

For \code{sync.cluster}, use \code{\link{sync_cluster}}.
}

\section{\code{sync.test}}{

For \code{sync.test}, use \code{\link{sync_test}}.
}

\section{\code{wavk.test}}{

For \code{wavk.test}, use \code{\link{wavk_test}}.
}

\author{
Calvin Chu, Yulia R. Gel, Vyacheslav Lyubchich

Vyacheslav Lyubchich

Vyacheslav Lyubchich, Yulia R. Gel

Srishti Vishwakarma, Vyacheslav Lyubchich

Yulia R. Gel, Vyacheslav Lyubchich, Ethan Schaeffer, Xingyu Wang

Yulia R. Gel, Vyacheslav Lyubchich, Ethan Schaeffer
}
\keyword{internal}
