% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notrend.test.R
\name{notrend.test-defunct}
\alias{notrend.test-defunct}
\title{Sieve Bootstrap Based Test for the Null Hypothesis of no Trend}
\description{
Sieve Bootstrap Based Test for the Null Hypothesis of no Trend
}
\seealso{
\code{\link{funtimes-defunct}}
}
\keyword{internal}
