% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.cluster.R
\name{sync.cluster-defunct}
\alias{sync.cluster-defunct}
\title{Time Series Clustering based on Trend Synchronism}
\description{
Time Series Clustering based on Trend Synchronism
}
\seealso{
\code{\link{funtimes-defunct}}
}
\keyword{internal}
