% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHomogeneity.R
\name{CHomogeneity}
\alias{CHomogeneity}
\title{Time Series Cluster Homogeneity}
\usage{
CHomogeneity(Bu, Bv, Alpha, Beta, Delta)
}
\arguments{
\item{Bu}{bucket of time series already included in the cluster.}

\item{Bv}{bucket of time series (neighbors) for potential inclusion in the cluster.}

\item{Alpha}{lower limit of the time series domain.}

\item{Beta}{upper limit of the time series domain.}

\item{Delta}{closeness parameter, a real value in \eqn{[0,1]}.}
}
\value{
A logical value indicating whether time series in \code{Bu} and \code{Bv} 
form a homogeneous cluster.
}
\description{
This is an auxiliary function to check homogeneity of time series cluster, based on 
Definition 4 by \insertCite{Ciampi_etal_2010;textual}{funtimes}.
}
\examples{
Bu <- rnorm(10)
Bv <- rnorm(10)
Alpha <- min(c(Bu, Bv))
Beta <- max(c(Bu, Bv))
CHomogeneity(Bu, Bv, Alpha, Beta, Delta = 0.5)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{CSlideCluster}}, \code{\link{CWindowCluster}}, 
and \code{\link{BICC}}
}
\author{
Vyacheslav Lyubchich
}
\keyword{cluster,}
\keyword{trend}
\keyword{ts,}
