% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.tails.R
\name{q.tails}
\alias{q.tails}
\title{Quantile-Based Tails Comparison}
\usage{
q.tails(x0, x1, q = 0.99)
}
\arguments{
\item{x0}{vectors of the same length (preferably). 
Tail in \code{x1} is compared against the tail in \code{x0}.}

\item{x1}{vectors of the same length (preferably). 
Tail in \code{x1} is compared against the tail in \code{x0}.}

\item{q}{a quantile defining the right tail for both \code{x0} and \code{x1}. 
Values above the  thresholds \code{quantile(x0, probs = q)} and 
\code{quantile(x1, probs = q)} are considered as the respective right tails.}
}
\value{
A list with two elements:
\item{d}{the quantile step.}
\item{Pk}{vector of differences of the intervals' centers.}
}
\description{
This function compares right tails of two sample distributions using 
a quantile-based approach (QBA); 
see \insertCite{Soliman_etal_2014_insurance;textual}{funtimes},
\insertCite{Soliman_etal_2015_insurance;textual}{funtimes}, 
and \insertCite{Lyubchich_Gel_2017_insurance;textual}{funtimes}.
}
\details{
Sturges' formula is used to calculate number of intervals (\eqn{k}) 
to split the upper \eqn{100(1 - q)}\% portion of \code{x0} and \code{x1} 
(the right tails). Then, each tail is divided into equally-filled intervals 
with a quantile step \eqn{d=(1 - q)/k}. \code{Pk} reports the difference between 
corresponding intervals' centers obtained from \code{x0} and \code{x1}.
}
\examples{
x0 <- rnorm(1000)
x1 <- rt(1000, 5)
q.tails(x0, x1)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{i.tails}}
}
\author{
Vyacheslav Lyubchich, Yulia R. Gel
}
\keyword{ts}
