% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.cluster.R
\name{sync.cluster}
\alias{sync.cluster}
\title{Time Series Clustering based on Trend Synchronism}
\usage{
sync.cluster(formula, rate = 1, alpha = 0.05, ...)
}
\arguments{
\item{formula}{an object of class "\code{\link[stats]{formula}}", 
specifying the type of common trend 
for clustering the time series in a \eqn{T} by \eqn{N} matrix of time series 
(time series in columns) which is passed to \code{\link{sync.test}}. 
Variable \eqn{t} should be used to specify the form 
of the trend, where \eqn{t} is specified within the function automatically as a 
regular sequence of length \eqn{T} on the interval (0,1]. See `Examples'.}

\item{rate}{rate of removal of time series. Default is 1 (i.e., if hypothesis 
of synchronism is rejected one time series is removed at a time to re-test the 
remaining time series). Integer values above 1 are treated as number of time 
series to be removed. Values from 0 to 1 are treated as a fraction of 
time series to be removed.}

\item{alpha}{significance level for testing hypothesis of a common trend 
(using \code{\link{sync.test}}) of the parametric form specified in \code{formula}.}

\item{...}{arguments to be passed to \code{\link{sync.test}}, for example, 
number of bootstrap replications (\code{B}).}
}
\value{
A list with the elements:
 #SL: need to make the output more user-friendly. Where are the clustering results (cluster assignments or so-called cluster labels)? #SL: Also, please, use similar notations to other functions.
\item{clusters}{total number of clusters obtained.}
\item{cluster_label}{cluster labels of each time series (each label is the cluster number of 
time series). A label zero represents the time series which does not have a common trend with
other time series. Labels other than zero are the clusters obtained.}
\item{column_index}{stores index of columns of time series in the main matrix corresponding to each cluster. This can be used for
plotting the time series of clustered time series, for example.}
\item{estimate}{parametric trend estimates obtained for each cluster.}
\item{pval}{\code{p}-value of \code{sync.test} for each cluster.}
\item{statistics}{value of \code{sync.test} test statistics for each cluster.}
\item{ar_order}{AR filter order used in \code{sync.test} for each cluster.}
\item{window_used}{window used in the \code{sync.test} test for each cluster.}
\item{all_considered_Windows}{different windows considered in \code{sync.test} for each cluster.}
\item{WAVK_obs}{WAVK test statistics for each cluster obtained from \code{sync.test} for each cluster.}
In case of more than one cluster, each output will be in 
the same sequence as of unique cluster labels except zero.
}
\description{
Cluster time series with a common parametric trend using the 
\code{\link{sync.test}} function 
\insertCite{Lyubchich_Gel_2016_synchronism,Ghahari_etal_2017_MBDCE}{funtimes}.
}
\details{
The \code{sync.cluster} function recursively clusters time series having 
a pre-specified common parametric trend until there are no time series left. 
Starting with the given \eqn{N} time series, the \code{\link{sync.test}} function 
is used to test for a common trend. If null hypothesis of common trend is not 
rejected by \code{\link{sync.test}}, the time series are grouped together 
(i.e., assigned to a cluster). Otherwise, the time series with the largest 
contribution to the test statistics are temporarily removed (the number of time 
series to remove depends on the \code{rate} of removal) and \code{\link{sync.test}} 
is applied again. The contribution to the test statistic is assessed by the
WAVK test statistic calculated for each time series.
}
\examples{
## Simulate four autoregressive time series, 
## three having a linear trend and one without a trend:
set.seed(123)
T = 30
N = 4
X = sapply(1:N, function(x) arima.sim(n = T + 100, list(order = c(1, 0, 0), ar = c(0.6)))[-c(1:100)])
X[,1] <- 5 * (1:T)/T + X[,1]
plot.ts(X)

# Finding clusters with default parameters
\dontrun{
    LinTrend <- sync.cluster(X ~ t) 
}    
## Sample Output:
## "Number of Clusters obtained: 1"
## "Cluster Labels"
## 1 1 0 0

## plotting the time series of the cluster obtained
idx = LinTrend$column_index
plot.ts(X[,idx])

## simulating seven autoregressive time series 
## Four time series have linear trend added 
set.seed(123)
T = 30
N = 10
X = matrix(NA, nrow = T, ncol = N)
X[,1:5] <-  sapply(1:5, function(x) arima.sim(n = T + 100, list(order = c(1, 0, 0), ar = c(0.6)))[-c(1:100)])
X[,6:N] <- sapply(6:N, function(x) -10 + 0.5 * (1:T) + arima.sim(n = T + 100, list(order = c(1, 0, 0), ar = c(0.6)))[-c(1:100)])
plot.ts(X)

## Clustering with rate of removal = 5, and window = 15
\dontrun{
    LinTrendR5W15 <- sync.cluster(X~t, rate = 2, Window = 15) 
}    
## Sample output:
## "Number of Clusters obtained: 2"
## "Cluster Labels"
##   2 1 1 2 1 1 1 1 1 1

## simulating five autoregressive time series to test for quadratic trend
## One has no trend, while rest of the series have quadratic trend
set.seed(123)
T = 30
N = 5
X = matrix(NA, nrow = T, ncol = N)
p <- 0.5
q <- 1:T

X[,1:2] <-  sapply(1:2, function(x) arima.sim(n = T + 100, list(order = c(1, 0, 0), ar = c(0.6)))[-c(1:100)])
X[,3:N] <- sapply(3:N, function(x) -10 + p*(q+10)^2 +arima.sim(n = T + 100, list(order = c(1, 0, 0), ar = c(0.6)))[-c(1:100)])
  
plot.ts(X)
# Clustering with default rate of removal
\dontrun{
    QuadTrend <- sync.cluster(X~poly(t,2))
}
## Sample output:
## "Number of Clusters obtained: 1"
## "Cluster Labels"
## 1 1 0 1 1  


}
\references{
\insertAllCited{}
}
\author{
Srishti Vishwakarma, Vyacheslav Lyubchich
}
\keyword{cluster}
\keyword{synchrony}
\keyword{trend}
