% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.cluster.R
\name{sync.cluster}
\alias{sync.cluster}
\title{Time Series Clustering based on Trend Synchronism}
\usage{
sync.cluster(formula, rate = 1, alpha = 0.05, ...)
}
\arguments{
\item{formula}{an object of class "\code{\link[stats]{formula}}", 
specifying the type of common trend 
for clustering the time series in a \eqn{T} by \eqn{N} matrix of time series 
(time series in columns). It is passed to \code{\link{sync.test}}. 
Variable \eqn{t} should be used to specify the form 
of the trend, where \eqn{t} is specified within the function automatically as a 
regular sequence of length \eqn{T} on the interval (0,1]. See `Examples'.}

\item{rate}{rate of removal of time series. Default is 1 (i.e., if hypothesis 
of synchronism is rejected one time series is removed at a time to re-test the 
remaining time series). Integer values above 1 are treated as number of time 
series to be removed. Values from 0 to 1 are treated as a fraction of 
time series to be removed.}

\item{alpha}{significance level for testing hypothesis of a common trend 
(using \code{\link{sync.test}}) of the parametric form specified in \code{formula}.}

\item{...}{arguments to be passed to \code{\link{sync.test}}, for example, 
number of bootstrap replications (\code{B}).}
}
\value{
A list with the elements: #SL: need to make the output more user-friendly. Where are the clustering results (cluster assignments or so-called cluster labels)? #SL: Also, please, use similar notations to other functions.
\item{Clusters}{number of clusters obtained.}
\item{Column.Index}{index of columns of time series (based on the main matrix) in each cluster.}
\item{Estimate}{parametric trend estimates of clusters obtained.}
\item{Pval}{\code{p}-value of the \code{sync.test}.}
\item{Statistics}{value of \code{sync.test} test statistics.}
\item{ar.order}{AR filter order.}
\item{window_used}{window used for the \code{sync.test} test.}
\item{all_considered_Windows}{different windows considered in the \code{sync.test}.}
\item{WAVK_obs}{WAVK test statistics for each cluster obtained from \code{sync.test}.}
}
\description{
Cluster time series with a common parametric trend using the 
\code{\link{sync.test}} function 
\insertCite{Lyubchich_Gel_2016_synchronism,Ghahari_etal_2017_MBDCE}{funtimes}.
}
\details{
The \code{sync.cluster} function recursively clusters time series having 
a pre-specified common parametric trend until there are no time series left. 
Starting with the given \eqn{N} time series, the \code{\link{sync.test}} function 
is used to test for a common trend. If null hypothesis of common trend is not 
rejected by \code{\link{sync.test}}, the time series are grouped together 
(i.e., assigned to a cluster). Othewise, the time series with the largest 
contribution to the test statistics are temporarily removed (the number of time 
series to remove depends on the \code{rate} of removal) and \code{\link{sync.test}} 
is applied again. The contribution to the test statistic is assessed by the
WAVK test statistic calculated for each time series.
}
\examples{
## Simulate four autoregressive time series, 
## three having a linear trend and one without a trend:
set.seed(123)
T = 100
N = 4
X = sapply(1:N, function(x) arima.sim(n = T + 100, list(order = c(1, 0, 0), ar = c(0.6)))[-c(1:100)])
X[,1] <- 5 * (1:T)/T + X[,1]
plot.ts(X)

# Finding clusters with default parameters
\dontrun{
    LinTrend <- sync.cluster(X ~ t) 
}    
## Sample Output:
## $`Clusters`
## Lfinal
## 1 
## 3 

## simulating seven auroregressive time series 
## Three have linear trend added and four have no  trend
n = 50
nc = 7
Y = matrix(NA, nrow = n, ncol = nc)
for ( i in 1:nc){
    if (i < 5){
        Y[,i] <-  arima.sim(n = n, list(order = c(1, 0, 0), ar = c(0.6)))
    } else {
        Y[,i] <- -10 + 0.5 * (1:n) + arima.sim(n = n, list(order = c(1, 0, 0), ar = c(0.6)))
    }
}

plot.ts(Y)
## Clustering with rate of removal = 5 and window = 15
\dontrun{
    LinTrendR5W15 <- sync.cluster(Y~t, rate = 5, Window = 15) 
}    
## Sample output:
# $`Clusters`
# Lfinal
# 1 
# 2 


## simulating five autoregressive time series to test for quadratic trend
## One has no trend, while rest of the series have quadratic trend
n = 30
nc = 5
Y = matrix(NA, nrow = n, ncol = nc)
p <- 0.5
q <- 1:n

for ( i in 1:nc){
    if (i < 2){
        Y[,i] <- arima.sim(n = n, list(order = c(1, 0, 0), ar = c(0.6)))
    } else {
        Y[,i] <- -10 + p*(q+10)^2 + arima.sim(n = n, list(order = c(1, 0, 0), ar = c(0.6)))
    }
    
}
plot.ts(Y)
# Clustering with default rate of removal
\dontrun{
    QuadTrend <- sync.cluster(Y~poly(t,2))
}
## Sample output:
## $`Clusters`
## Lfinal
## 0 1 
## 2 2 
    
}
\references{
\insertAllCited{}
}
\author{
Srishti Vishwakarma, Vyacheslav Lyubchich
}
\keyword{cluster}
\keyword{synchrony}
\keyword{trend}
